feat: Reorganizer events browser
