+++ lib/firebase_messaging_background_handler.dart
@@ -31,8 +31,8 @@ import 'package:bluecherry_client/providers/settings_provider.dart';
 import 'package:bluecherry_client/utils/constants.dart';
 import 'package:bluecherry_client/utils/storage.dart';
 import 'package:bluecherry_client/utils/video_player.dart';
-import 'package:bluecherry_client/widgets/events/events_screen.dart';
-import 'package:bluecherry_client/widgets/player/live_player.dart';
+import 'package:bluecherry_client/screens/events_browser/events_screen.dart';
+import 'package:bluecherry_client/screens/players/live_player.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
 import 'package:flutter/material.dart';
