+++ lib/main.dart
@@ -45,12 +45,12 @@ import 'package:bluecherry_client/utils/theme.dart';
 import 'package:bluecherry_client/utils/video_player.dart';
 import 'package:bluecherry_client/utils/window.dart';
 import 'package:bluecherry_client/widgets/desktop_buttons.dart';
-import 'package:bluecherry_client/widgets/events/events_screen.dart';
+import 'package:bluecherry_client/screens/events_browser/events_screen.dart';
 import 'package:bluecherry_client/widgets/home.dart';
 import 'package:bluecherry_client/widgets/multi_window/single_camera_window.dart';
 import 'package:bluecherry_client/widgets/multi_window/single_layout_window.dart';
 import 'package:bluecherry_client/widgets/multi_window/window.dart';
-import 'package:bluecherry_client/widgets/player/live_player.dart';
+import 'package:bluecherry_client/screens/players/live_player.dart';
 import 'package:bluecherry_client/widgets/splash_screen.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
