+++ lib/screens/events_browser/events_screen_desktop.dart
