+++ lib/screens/players/live_player.dart
@@ -32,7 +32,6 @@ import 'package:bluecherry_client/widgets/error_warning.dart';
 import 'package:bluecherry_client/widgets/hover_button.dart';
 import 'package:bluecherry_client/widgets/misc.dart';
 import 'package:bluecherry_client/widgets/multi_window/window.dart';
-import 'package:bluecherry_client/widgets/player/widgets.dart';
 import 'package:bluecherry_client/widgets/ptz.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/gestures.dart';
